+++ include/CKParameterOut.h
@@ -50,7 +50,7 @@ public:
     // Value
 
     virtual CKERROR GetValue(void *buf, CKBOOL update = TRUE);
-    virtual CKERROR SetValue(const void *buf, int size = 0);
+    virtual CKERROR SetValue(const void *buf, size_t size = 0);
     virtual CKERROR CopyValue(CKParameter *param, CKBOOL UpdateParam = TRUE);
     virtual void *GetReadDataPtr(CKBOOL update = TRUE);
     virtual int GetStringValue(CKSTRING Value, CKBOOL update = TRUE);
